// whoxa_frontend/src/types/ResType.ts
export interface ConversationDetails {
  group_profile_image: string;
  group_name: string;
}
// {{base_url}}create-group response ==================================================================================

// Update Group Response ==================================================================================
export interface UpdateGroupRes {
  success: boolean;
  message: string;
  conversation_id: string;
  conversationDetails: ConversationDetails;
}